// src/client/types.gen.ts
export type ProductRead = {
    /**
     * Id
     */
    id: string;
    /**
     * Title
     */
    title: string;
    /**
     * Price
     */
    price: string;
    status: ProductStatus;
    /**
     * Created At
     */
    created_at: string;
};

/**
 * ProductStatus
 */
export type ProductStatus = 'AVAILABLE' | 'RESERVED' | 'SOLD';